server/requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
flask-cors==3.0.6
psycopg2==2.7.3
sqlalchemy==1.2.7
sqlalchemy-migrate==0.11.0
SQLAlchemy-Utils==0.32.16
flask-cors==3.0.6
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.14.1
pymysql==0.9.3